--- src/Components/Home/Inicio.css
@@ -1,40 +1,0 @@
-/* Inicio.css */
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-  background-color: #f5f5f5;
-}
-
-.header {
-  font-size: 2rem;
-  font-weight: bold;
-  margin-bottom: 20px;
-  color: #333;
-}
-
-.description {
-  font-size: 1.2rem;
-  color: #666;
-  text-align: center;
-  max-width: 600px;
-  margin-bottom: 30px;
-}
-
-.button {
-  padding: 10px 20px;
-  font-size: 1rem;
-  color: #fff;
-  background-color: #007bff;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.button:hover {
-  background-color: #0056b3;
-}
