+++ src/Components/Navegation/MenuAppBar.jsx
@@ -1,124 +1,68 @@
-import React, { useState } from "react";
+import React from "react";
 import AppBar from "@mui/material/AppBar";
 import Toolbar from "@mui/material/Toolbar";
 import Typography from "@mui/material/Typography";
 import IconButton from "@mui/material/IconButton";
 import MenuIcon from "@mui/icons-material/Menu";
 import SearchIcon from "@mui/icons-material/Search";
-import InputBase from "@mui/material/InputBase";
-import { styled, alpha } from "@mui/material/styles";
 import Button from "@mui/material/Button";
 import { Link } from "react-router-dom";
 import Brightness4Icon from "@mui/icons-material/Brightness4";
 import Brightness7Icon from "@mui/icons-material/Brightness7";
-import { ThemeProvider, createTheme } from "@mui/material/styles";
 import "./MenuAppBar.css";
+import { Search, SearchIconWrapper, StyledInputBase } from "./StylesAppBar";
 
-const Search = styled("div")(({ theme }) => ({
-  position: "relative",
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  "&:hover": {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginRight: theme.spacing(2),
-  marginLeft: 0,
-  width: "100%",
-  [theme.breakpoints.up("sm")]: {
-    marginLeft: theme.spacing(3),
-    width: "auto",
-  },
-}));
-
-const SearchIconWrapper = styled("div")(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: "100%",
-  position: "absolute",
-  pointerEvents: "none",
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-}));
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: "inherit",
-  "& .MuiInputBase-input": {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create("width"),
-    width: "100%",
-    [theme.breakpoints.up("md")]: {
-      width: "20ch",
-    },
-  },
-}));
-
-const MenuAppBar = () => {
-  const [darkMode, setDarkMode] = useState(false);
-  //   const theme = useTheme();
-  const darkTheme = createTheme({
-    palette: {
-      mode: darkMode ? "dark" : "light",
-    },
-  });
-
-  const handleThemeChange = () => {
-    setDarkMode(!darkMode);
-  };
-
+const MenuAppBar = ({ mode, toggleTheme }) => {
   return (
-    <ThemeProvider theme={darkTheme}>
-      <AppBar position="static">
-        <Toolbar>
-          <IconButton
-            edge="start"
-            color="inherit"
-            aria-label="open drawer"
-            sx={{ mr: 2 }}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Typography
-            variant="h6"
-            noWrap
-            component="div"
-            sx={{ flexGrow: 1, display: { xs: "none", sm: "block" } }}
-          >
-            LOGO
-          </Typography>
-          <Search>
-            <SearchIconWrapper>
-              <SearchIcon />
-            </SearchIconWrapper>
-            <StyledInputBase
-              placeholder="Buscar eventos"
-              inputProps={{ "aria-label": "search" }}
-            />
-          </Search>
-          <Button color="inherit">
-            <Link className="Link" to="/" color="inherit">
-              Inicio
-            </Link>
-          </Button>
-          <Button color="inherit">
-            <Link className="Link" to="/eventosprivados">
-              Eventos
-            </Link>
-          </Button>
-          <Button color="inherit">Conócenos</Button>
-          <Button color="inherit">Centro de Ayuda</Button>
-          <Button color="inherit">
-            <Link className="Link" to="/login">
-              Login
-            </Link>
-          </Button>
-          <IconButton color="inherit" onClick={handleThemeChange}>
-            {darkMode ? <Brightness7Icon /> : <Brightness4Icon />}
-          </IconButton>
-        </Toolbar>
-      </AppBar>
-    </ThemeProvider>
+    <AppBar position="static">
+      <Toolbar>
+        <IconButton
+          edge="start"
+          color="inherit"
+          aria-label="open drawer"
+          sx={{ mr: 2 }}
+        >
+          <MenuIcon />
+        </IconButton>
+        <Typography
+          variant="h6"
+          noWrap
+          component="div"
+          sx={{ flexGrow: 1, display: { xs: "none", sm: "block" } }}
+        >
+          LOGO
+        </Typography>
+        <Search>
+          <SearchIconWrapper>
+            <SearchIcon />
+          </SearchIconWrapper>
+          <StyledInputBase
+            placeholder="Buscar eventos"
+            inputProps={{ "aria-label": "search" }}
+          />
+        </Search>
+        <Button color="inherit">
+          <Link className="Link" to="/" color="inherit">
+            Inicio
+          </Link>
+        </Button>
+        <Button color="inherit">
+          <Link className="Link" to="/eventosprivados">
+            Eventos
+          </Link>
+        </Button>
+        <Button color="inherit">Conócenos</Button>
+        <Button color="inherit">Centro de Ayuda</Button>
+        <Button color="inherit">
+          <Link className="Link" to="/login">
+            Login
+          </Link>
+        </Button>
+        <IconButton color="inherit" onClick={toggleTheme}>
+          {mode === "dark" ? <Brightness7Icon /> : <Brightness4Icon />}
+        </IconButton>
+      </Toolbar>
+    </AppBar>
   );
 };
