+++ src/Components/Security/Login.jsx
@@ -11,87 +11,82 @@ import {
 } from "@mui/material";
 import { Link } from "react-router-dom";
 import PersonIcon from "@mui/icons-material/Person";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
-
-const theme = createTheme();
 
 const Login = () => {
   return (
-    <ThemeProvider theme={theme}>
-      <Container
-        component="main"
-        maxWidth="xs"
+    <Container
+      component="main"
+      maxWidth="xs"
+      sx={{
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "center",
+        height: "100vh",
+      }}
+    >
+      <CssBaseline />
+      <Box
         sx={{
           display: "flex",
           flexDirection: "column",
-          justifyContent: "center",
-          height: "100vh",
+          alignItems: "center",
         }}
       >
-        <CssBaseline />
-        <Box
-          sx={{
-            display: "flex",
-            flexDirection: "column",
-            alignItems: "center",
-          }}
-        >
-          <Avatar sx={{ m: 1, bgcolor: "primary.main" }}>
-            <PersonIcon />
-          </Avatar>
-          <Typography component="h1" variant="h5">
-            Inicia Sesión
-          </Typography>
-          <Box component="form" noValidate sx={{ mt: 1 }}>
-            <TextField
-              margin="normal"
-              required
-              fullWidth
-              id="email"
-              label="Correo"
-              name="email"
-              autoComplete="email"
-              autoFocus
-            />
-            <TextField
-              margin="normal"
-              required
-              fullWidth
-              name="password"
-              label="Contraseña"
-              type="password"
-              id="password"
-              autoComplete="current-password"
-            />
-            <Button
-              type="submit"
-              fullWidth
-              variant="contained"
-              sx={{ mt: 3, mb: 2 }}
-            >
-              Iniciar Sesión
-            </Button>
-            <Grid container>
-              <Grid item>
-                <Link to="/" variant="body2">
-                  {"¿Olvidaste tu contraseña?"}
-                </Link>
-              </Grid>
-            </Grid>
-            <Button
-              type="submit"
-              fullWidth
-              variant="contained"
-              sx={{ mt: 3, mb: 2 }}
-            >
-              <Link className="Link" to="/registro">
-                Registrarse
+        <Avatar sx={{ m: 1, bgcolor: "primary.main" }}>
+          <PersonIcon />
+        </Avatar>
+        <Typography component="h1" variant="h5">
+          Inicia Sesión
+        </Typography>
+        <Box component="form" noValidate sx={{ mt: 1 }}>
+          <TextField
+            margin="normal"
+            required
+            fullWidth
+            id="email"
+            label="Correo"
+            name="email"
+            autoComplete="email"
+            autoFocus
+          />
+          <TextField
+            margin="normal"
+            required
+            fullWidth
+            name="password"
+            label="Contraseña"
+            type="password"
+            id="password"
+            autoComplete="current-password"
+          />
+          <Button
+            type="submit"
+            fullWidth
+            variant="contained"
+            sx={{ mt: 3, mb: 2 }}
+          >
+            Iniciar Sesión
+          </Button>
+          <Grid container>
+            <Grid item>
+              <Link to="/" variant="body2">
+                {"¿Olvidaste tu contraseña?"}
               </Link>
-            </Button>
-          </Box>
+            </Grid>
+          </Grid>
+          <Button
+            type="submit"
+            fullWidth
+            variant="contained"
+            sx={{ mt: 3, mb: 2 }}
+          >
+            <Link className="Link" to="/registro">
+              Registrarse
+            </Link>
+          </Button>
         </Box>
-      </Container>
-    </ThemeProvider>
+      </Box>
+    </Container>
   );
 };
